tests: split www/companies_views/test_views.py file
